Refactor UserService to use UserDTO for user conversion
This commit refactors the UserService class in the src/main/java/falusvampen/letsplay/service/UserService.java file. It introduces a new import for UserDTO and adds a conversion method, convertToUserDTO, to convert User objects to UserDTO objects. The getAllUsers and getUserById methods are then modified to use this conversion method and return a List of UserDTO objects instead of User objects.